Move initialisation of The_Image to proving
Parameter The_Image should always be initialised as it's based on the
size. We shall see what we can do at proving time to make SPARK happier
in the future, though it partly depends on instantiations. We could make
it in out and allow the user to or initailse it to 'U' for unset but it
would be a performance issue and the intention is that it should always
be the right length.